refactor(workspace): split workspace.rs into domain submodules
